Refactor vpn indicator into separate file
